meeting_assistant/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
aiohttp>=3.9.0
aiofiles>=23.2.0
zhipuai>=2.0.0
python-dotenv>=1.0.0
sniffio>=1.3.0
python-docx>=1.1.0
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

